HUUUUUGE UPDATE
Added spaceship model with player controls.  Press Q/E to rotate your
ship and WASD to move around.  Press Space to fire bullets.  Music and
bullet SFX added.  Bullet design is currently a placeholder.  Added
skybox space background and pause menu.  Press P to pause.  This will
pause the game and bring up a pause display, as well as pausing all
active audio.  Other things too that are too lengthy to explain in here,
BUT WE GOTS OURSELVES THE BASIS TO A GAME!!!